Creo estructura de carrito, agrego boton para eliminar pedido y volver al menu
