+++ package.json
@@ -15,19 +15,19 @@
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.15.6",
     "@mui/material": "^5.15.6",
-    "@tanstack/react-query": "^5.17.19",
+    "@tanstack/react-query": "^5.37.1",
     "bootstrap": "^5.3.2",
     "bootstrap-icons": "^1.11.3",
     "classnames": "^2.5.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.49.3",
+    "react-hook-form": "^7.51.5",
     "react-icons": "^5.0.1",
-    "react-router-dom": "^6.21.3",
-    "sonner": "^1.3.1",
-    "sweetalert2": "^11.10.4",
-    "zustand": "^4.5.0"
+    "react-router-dom": "^6.23.1",
+    "sonner": "^1.4.41",
+    "sweetalert2": "^11.11.0",
+    "zustand": "^4.5.2"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
