+++ src/components/Menu/Menu.css
@@ -55,7 +55,7 @@
 }
 
 .menuImage {
-  width: 200px;
+  width: 100px;
 }
 
 .menuImage img {
